reconfiguration to get working version of basic files
