+++ JARVIS/.classpath
@@ -2,13 +2,13 @@
 <classpath>
 	<classpathentry kind="src" path="src"/>
 	<classpathentry exported="true" kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry kind="lib" path="src/cmu_us_kal.jar"/>
-	<classpathentry kind="lib" path="src/cmulex.jar"/>
-	<classpathentry kind="lib" path="src/en_us.jar"/>
-	<classpathentry kind="lib" path="src/freetts.jar"/>
-	<classpathentry kind="lib" path="src/jsapi.jar"/>
-	<classpathentry kind="lib" path="src/sphinx4.jar"/>
-	<classpathentry kind="lib" path="src/voce.jar"/>
-	<classpathentry kind="lib" path="src/WSJ_8gau_13dCep_16k_40mel_130Hz_6800Hz.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/cmu_us_kal.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/cmulex.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/en_us.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/freetts.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/jsapi.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/sphinx4.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/voce.jar"/>
+	<classpathentry kind="lib" path="/Users/austinhulse/Documents/workspace/ButlerJarvis/JARVIS-BUTLER/JARVIS/lib/WSJ_8gau_13dCep_16k_40mel_130Hz_6800Hz.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
