+++ JARVIS/lib/grammar/fruit.gram
@@ -1,0 +1,11 @@
+#JSGF V1.0;
+
+/**
+ * JSGF Digits Grammar for Hello World example
+ */
+
+grammar fruit;
+
+public <fruitTypes> = (apple | orange | banana);
+
+public <commandTypes> = (quit);
